You can get an image from an URL in a region using diaspora-show-image-at-region.
This function was created because diaspora-show-image-at-point not always succeeds in finding the URL of the actual image. And what's happens if the user only give an URL instead of the markdown? It fails as well and you should use diaspora-show-image-at-region.

Image-paths in temporary directory can be obtained by its URL or by the filename.